requirements.txt
attrs==23.2.0
blinker==1.7.0
blueprint==3.4.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.0.2
flask-marshmallow==1.2.1
Flask-MySQL==1.5.2
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
greenlet==3.0.3
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
jsonify==0.5
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
linecache2==1.0.0
load-dotenv==0.1.0
MarkupSafe==2.1.5
marshmallow==3.21.1
marshmallow-sqlalchemy==1.0.0
mistune==3.0.2
mysqlclient==2.2.4
packaging==24.0
pluggy==1.4.0
PyMySQL==1.1.0
pytest==8.1.1
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rpds-py==0.18.0
six==1.16.0
SQLAlchemy==2.0.28
traceback2==1.4.0
typing_extensions==4.10.0
unittest2==1.1.0
urllib3==2.2.1
Werkzeug==3.0.1
